// frontend/proAngular/src/app/models/sadnica.ts
export class SadnicaPost{
    naziv: string;
    progres: number;
    vek: number;
    brPolja: number;
    uklonjena: boolean;
    proizvodjac: string;

}


export class Sadnica{
    id: number;
    naziv: string;
    progres: number;
    vek: number;
    brPolja: number;
    uklonjena: boolean;
    proizvodjac: string;
    
}